requirements.txt
opencv-python-headless
numpy
pandas
pytesseract
streamlit
Pillow
